feat: componentized dialog action confirmation delete course

+++ crud-angular-spring/crud-angular/src/app/courses/container/courses/courses.component.ts
@@ -6,6 +6,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -37,7 +38,7 @@ export class CoursesComponent {
 
   refresh() {
     this.courses$ = this.coursesService?.list().pipe(
-      catchError( error => {
+      catchError(error => {
         this.onError('Erro ao carregar cursos')
         return of([])
       })
@@ -51,21 +52,24 @@ export class CoursesComponent {
   }
 
   onDelete(course: Course) {
-    this.coursesService.delete(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso Removido com sucesso!', 'X',
-          {
-            duration: 5000,
-            verticalPosition: 'top',
-            horizontalPosition: 'center'
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?'
+    })
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.delete(course._id).subscribe({
+          next: () => {
+            this.refresh();
+            this.snackBar.open('Curso Removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
           },
-        )
-      },
-      () => this.onError('Erro ao tentar remover curso.')
-
-    );
+          error: () => this.onError('Erro ao tentar remover curso.')
+        });
+      }
+    })
   }
-
-
 }
